Rename communicability.py to communicability_alg.py
